So I was thinking of a way to get the synchronization working nicely, thinking of WaitFor( n ) and Sync( tid ) when a couple o' guys that were up to no good, started makin' trouble in the neighbourhood. I got in one little fight and my mom got scared, She said 'You're movin with your auntie and uncle in Bel-Air!'
So, I whistled for a cab and when i came near,
The license plate said 'Fresh',
And had dice in the mirror,
If anything i could say that this cab was rare,
But I thought 'Nah, forget it - Yo, home to Bel-Air!'

I pulled up to the house at bout seven or eight,
I yelled to the cabbie 'Yo home, smell ya later!'
I looked at my kingdom,
I was finally there!
To sit on my throne as the prince of Bel-Air!